+++ structs/basic_inventory_assessments.struct.go
@@ -7,8 +7,8 @@ type BasicInventoryAssessmentsTypesItem struct {
 	Active               bool    `json:"active"`
 	DepreciationTypeId   int     `json:"depreciation_type_id"`
 	UserProfileId        int     `json:"user_profile_id"`
-	GrossPriceNew        int     `json:"gross_price_new"`
-	GrossPriceDifference int     `json:"gross_price_difference"`
+	GrossPriceNew        float32 `json:"gross_price_new"`
+	GrossPriceDifference float32 `json:"gross_price_difference"`
 	DateOfAssessment     *string `json:"date_of_assessment"`
 	CreatedAt            string  `json:"created_at"`
 	UpdatedAt            string  `json:"updated_at"`
